refactor: extract shared runSubagentProcess + split taskflow-mcp into a 6th package
Two architectural cleanups from the cross-adversarial review's biggest finding
(the three host runners duplicated ~246 lines of identical process boilerplate,
which had already caused a divergence bug):

1. runSubagentProcess — eliminate the runner duplication.
   The codex/claude/opencode runners each copy-pasted the identical
   spawn / idle-watchdog / abort / signal-kill / stderr-cap / post-exit
   classification block. Extracted once into taskflow-core's runner-core.ts as
   runSubagentProcess (parameterized by a per-host SubagentAccumulator + foldLine),
   plus a shared unknownAgentResult and a single centralized activeChildren set
   (one exit handler now reaches every host's children, not three separate sets).
   Each host runner shrank to only its host-specific bits:
     codex 366→217, claude 417→266, opencode 397→247 lines (~450 lines removed).
   Behavior identical (1045/1045 tests pass); a new host can no longer drift
   the process/classify contract.

2. taskflow-mcp — split the MCP server into its own package.
   Moved mcp/ (jsonrpc stdio transport + taskflow_* tool handlers + DAG svg/outline
   renderer) out of taskflow-core into a new taskflow-mcp package. taskflow-core is
   again purely the portable engine; the MCP presentation layer is an
   independently-publishable unit that depends on core. The host adapters now
   depend on taskflow-mcp and import via taskflow-mcp/server|jsonrpc; pi-taskflow
   is unaffected (never used the MCP server).

Wiring: 6th workspace, root build script, publish.yml (publish order
core→mcp→adapters + version check), copy-readme, AGENTS/README/RELEASE/CHANGELOG
updated for six packages. Verified: typecheck clean (modulo 3 preexisting
unrelated gate-score.test.ts overload warnings), 1045/1045 tests pass, all six
packages build, codex/claude/opencode network-free MCP stdio e2e all pass.
