+++ scripts/copy-readme.mjs
@@ -21,7 +21,7 @@ const here = dirname(fileURLToPath(import.meta.url));
 const repoRoot = join(here, "..");
 
 // Packages that are published to npm and should carry the README.
-const PUBLISHABLE = new Set(["taskflow-core", "pi-taskflow", "codex-taskflow", "claude-taskflow", "opencode-taskflow"]);
+const PUBLISHABLE = new Set(["taskflow-core", "taskflow-mcp", "pi-taskflow", "codex-taskflow", "claude-taskflow", "opencode-taskflow"]);
 
 const REPO = "heggria/taskflow";
 const BRANCH = "main";
